requirements.txt
Django~=5.1.6
django-ninja~=1.3.0